+++ shared/src/commonMain/kotlin/data/credentials/CredentialAttributeCategorization.kt
@@ -8,7 +8,6 @@ import at.asitplus.wallet.ehic.EhicScheme
 import at.asitplus.wallet.eupid.EuPidScheme
 import at.asitplus.wallet.eupidsdjwt.EuPidSdJwtScheme
 import at.asitplus.wallet.healthid.HealthIdScheme
-import at.asitplus.wallet.idaustria.IdAustriaScheme
 import at.asitplus.wallet.lib.data.ConstantIndex
 import at.asitplus.wallet.mdl.MobileDrivingLicenceScheme
 import at.asitplus.wallet.por.PowerOfRepresentationScheme
@@ -30,7 +29,6 @@ interface CredentialAttributeCategorization {
             scheme: ConstantIndex.CredentialScheme?,
             representation: ConstantIndex.CredentialRepresentation
         ): Template = when (scheme) {
-            is IdAustriaScheme -> IdAustriaCredentialAttributeCategorization
             is EuPidScheme -> EuPidCredentialAttributeCategorization
             is EuPidSdJwtScheme -> EuPidSdJwtCredentialAttributeCategorization
             is MobileDrivingLicenceScheme -> MobileDrivingLicenceCredentialAttributeCategorization